+++ src/api/pbx.ts
@@ -30,7 +30,6 @@ import {
   bundleIdentifier,
   fcmApplicationId,
   isAndroid,
-  isWeb,
   retryInterval,
 } from '#/config'
 import { embedApi } from '#/embed/embedApi'
@@ -1379,7 +1378,7 @@ export class PBX extends EventEmitter {
       command,
       service_id,
       application_id,
-      user_agent: isWeb ? navigator.userAgent : 'react-native',
+      user_agent: ctx.auth.getUserAgentRaw(),
       username,
       device_id,
       device_id_voip,